// packages/ny-uttaksplan/src/veilederInfo/types.ts
export type VeilederMessageType = 'normal' | 'info' | 'advarsel' | 'feil';

export interface VeilederMessage {
    contentIntlKey: string;
    type: VeilederMessageType;
    formatContentAsHTML?: boolean;
    titleIntlKey?: string;
    values?: any;
    periodeId?: string;
    skjulesIOppsummering?: boolean;
    avvikType?: string;
}